api-gateway/requirements.txt
# FastAPI Core
fastapi==0.110.0
uvicorn[standard]==0.27.0
pydantic[email]==2.6.1
pydantic-settings==2.2.1
email-validator==2.1.0

# HTTP Client for Go service communication
httpx==0.26.0

# Database (Async SQLAlchemy)
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
aiosqlite==0.19.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis for caching and rate limiting
redis==5.0.1
hiredis==2.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.7

# Background tasks
celery[redis]==5.3.6

# Monitoring & Logging
structlog==24.1.0
prometheus-client==0.19.0

# WebSocket Support
python-socketio[asyncio]==5.11.0
redis[hiredis]==5.0.1

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.23.5
pytest-cov==4.1.0

# Environment management
python-dotenv==1.0.1